Test spit into a deployable and non-deployable test
